requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile '.\requirements.in'
#
annotated-types==0.5.0
    # via pydantic
anyio==3.7.1
    # via
    #   fastapi
    #   starlette
    #   watchfiles
certifi==2023.7.22
    # via requests
cffi==1.16.0
    # via gevent
charset-normalizer==3.2.0
    # via requests
click==8.1.7
    # via uvicorn
colorama==0.4.6
    # via
    #   click
    #   uvicorn
exceptiongroup==1.1.3
    # via anyio
fastapi==0.103.2
    # via -r .\requirements.in
gevent==23.9.1
    # via grequests
greenlet==2.0.2
    # via gevent
grequests==0.7.0
    # via -r .\requirements.in
h11==0.14.0
    # via uvicorn
httptools==0.6.0
    # via uvicorn
idna==3.4
    # via
    #   anyio
    #   requests
psycopg2-binary==2.9.8
    # via -r .\requirements.in
pycparser==2.21
    # via cffi
pydantic==2.4.2
    # via fastapi
pydantic-core==2.10.1
    # via pydantic
python-dotenv==1.0.0
    # via uvicorn
pyyaml==6.0.1
    # via uvicorn
requests==2.31.0
    # via
    #   -r .\requirements.in
    #   grequests
sniffio==1.3.0
    # via anyio
starlette==0.27.0
    # via fastapi
typing-extensions==4.8.0
    # via
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   uvicorn
urllib3==2.0.5
    # via requests
uvicorn[standard]==0.23.2
    # via -r .\requirements.in
watchfiles==0.20.0
    # via uvicorn
websockets==11.0.3
    # via uvicorn
zope-event==5.0
    # via gevent
zope-interface==6.0
    # via gevent

# The following packages are considered to be unsafe in a requirements file:
# setuptools
